# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentação da API Despesas Cartão",
        "description": "Descrição da API para gerenciar despesas de cartão",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/expenses_users/": {
            "post": {
                "tags": [
                    "Despesas"
                ],
                "summary": "Insere na base de dados os registros da nova despesa.",
                "operationId": "ba89cf0424cd23313e225ae09169ea50",
                "responses": {
                    "200": {
                        "description": "Successfully Done!"
                    },
                    "500": {
                        "description": "Unexpected error!"
                    },
                    "424": {
                        "description": "Request error, validate the data and try again!"
                    },
                    "429": {
                        "description": "Undefined Card!"
                    },
                    "430": {
                        "description": "It is not possible to enter this expense because the ID card balance ('') is insufficient!"
                    },
                    "431": {
                        "description": "It is not possible to enter an expense on this ID card ('') because its balance is zero!"
                    }
                }
            }
        },
        "/api/expenses_users/{id_user}": {
            "get": {
                "tags": [
                    "Despesas"
                ],
                "summary": "Lista todas as despesas, somente do usuário selecionado.",
                "operationId": "ccaff020c96af79c84bb9c513854acb3",
                "responses": {
                    "420": {
                        "description": "No registration!"
                    },
                    "425": {
                        "description": "Undefined User!"
                    }
                }
            }
        },
        "/api/expenses_users/{id_expense}": {
            "put": {
                "tags": [
                    "Despesas"
                ],
                "summary": "Atualiza os dados somente da despesa selecionada já cadastrada.",
                "operationId": "800b4d39c8af065cd3a4a351f130311d",
                "responses": {
                    "200": {
                        "description": "Successfully Done!"
                    },
                    "500": {
                        "description": "Unexpected error!"
                    },
                    "424": {
                        "description": "Request error, validate the data and try again!"
                    },
                    "430": {
                        "description": "It is not possible to enter this expense because the ID card balance ('') is insufficient!"
                    },
                    "431": {
                        "description": "It is not possible to enter an expense on this ID card ('') because its balance is zero!"
                    },
                    "432": {
                        "description": "Undefined Expense!"
                    },
                    "433": {
                        "description": "This expense does not refer to this card ID!"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Despesas"
                ],
                "summary": "Exclui na base de dados os registros da despesa selecionada.",
                "operationId": "83d961026be1d7ffafd72bf77975430a",
                "responses": {
                    "200": {
                        "description": "Successfully Done!"
                    },
                    "500": {
                        "description": "Unexpected error!"
                    },
                    "424": {
                        "description": "Request error, validate the data and try again!"
                    },
                    "432": {
                        "description": "Undefined Expense!"
                    }
                }
            }
        },
...
    },
...
}